+++ src/error.rs
@@ -1,7 +1,7 @@
 use thiserror::Error;
 
 /// A specialized [`Result`] type for this crate.
-pub type Result<T> = std::result::Result<T, Error>;
+pub type Result<T, E = Error> = std::result::Result<T, E>;
 
 /// An extension trait for [`Result`]s.
 pub trait ResultExt<T> {